Fill in gaps in the profile
Sometimes there are gaps in the profile (probably because a piece of
code holds the GIL for a long time).  They distort the profile.  This
commit tries to compensate by generating fake samples for the missed
time.

Experiments show that the 1st sample taken after such a long gap is
close to the place where GIL was released (assuming my theory about GIL
causing those gaps are true), so let's use that stack frame for our fake
samples.  These fake samples are marked by adding '*' to the end, so you
can highlight them by searching for \*.